Fix bad flooring
The game's implemementation of Z-position display has a bit of an odd quirk that it uses to floor (rather than truncate) a negative number. The game will do this by subtracting 10 from the calculated z-value, so that a value such as -0.03 will display as -0.1.

However, this is not required (or desired!) for more precise displays. This change removes this unnecessary subtraction and instead floors properly.

Note that this will cause some odd display shenanigans, like a proper (precise) coordinate showing as -0.124 but the minimap saying -0.2. This is a more correct (and precise) value though, so I'd argue this change is okay.